Expand and clarify consitency/durability docs in store.wit
The existing docs are somewhat vague about how the "read your writes"
consistency model works in practice, so I've tried to make them more explicit.
Also, they don't mention durability at all, so I've added a section dedicated to
that.

Note that I've generally erred on the side of maximum portability across host
implementations at the expense of strong guarantees for the guest.  Based on
previous conversations, my understanding is that we _do_ want to support
implementations backed by eventually consistent distributed systems, and that
means portable guest code cannot assume a stronger consistency model than what
such systems can deliver.  Concretely, we must consider the scenario where a
host has a pool of connections to multiple replicas in such a system such that a
single component instance which opens the same bucket multiple times might get a
different replica (each with its own view of the state) each time.

If we feel the guarantees described in these docs are too weak, we can certainly
strengthen them at the expense of host implementation flexibility.
Alternatively, we could add new APIs for querying and/or controlling the
durability and consistency models provided by the implementation -- or even
allow the guest to statically declare that it requires some specific consistency
model by importing a specific interface corresponding to that model, analogous
to what we did with the `atomics` interface.

Regardless of what set of (non-)guarantees and features we settle on, my main
priority is to be as clear as possible about them so that application developers
are not caught by surprise.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>